backend/pyproject.toml
[project]
name = "doodlematcher"
version = "0.1.0"
description = ""
authors = [
    {name = "Maksym Dombrov",email = "maksymdombrov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.116.1,<0.117.0",
    "uvicorn[standard]>=0.35.0,<0.36.0", 
    "pydantic>=2.11.7,<3.0.0",
    "qdrant-client>=1.15.1,<2.0.0",
    "pillow>=11.3.0,<12.0.0",
    "requests>=2.31.0,<3.0.0",
    "numpy>=2.0.0,<3.0.0",
    "onnxruntime (>=1.22.1,<2.0.0)",
    "transformers (>=4.55.4,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

